Code Tests/flask_note/requirements.txt
gunicorn==20.0.4
celery==4.4.5
marshmallow==3.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.23.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.9
vine==1.3.0
Werkzeug==1.0.1
click==7.1.2
passlib==1.7.2
apispec==3.3.1
apispec-webframeworks==0.5.2
tox==3.15.2
marshmallow-sqlalchemy==0.23.1
Flask==1.1.2
Flask-Migrate==2.5.3
psycopg2-binary==2.8.5
python-dotenv==0.13.0
Flask-RESTful==0.3.8
flask-marshmallow==0.13.0
Flask-WTF==0.14.3
micawber==0.5.1
Markdown==3.2.2
pymongo==3.10.1
Flask-PyMongo==2.3.0